Revert "tagging: add iD Tagging scheme to advanced mode 🎉 (#131)"
This reverts commit 5a492a70
