+++ src/components/FeaturePanel/FeaturedTags.tsx
@@ -1,5 +1,7 @@
+import Typography from '@material-ui/core/Typography';
 import React from 'react';
 import styled from 'styled-components';
+import { t } from '../../services/intl';
 import { FeaturedTag } from './FeaturedTag';
 
 const Spacer = styled.div`
@@ -15,6 +17,10 @@ export const FeaturedTags = ({ featuredTags, setDialogOpenedWith }) => {
         <FeaturedTag key={k} k={k} v={v} onEdit={setDialogOpenedWith} />
       ))}
       <Spacer />
+
+      <Typography variant="overline" display="block" color="textSecondary">
+        {t('featurepanel.other_info_heading')}
+      </Typography>
     </>
   );
 };
