+++ src/services/fetchCache.ts
@@ -7,7 +7,6 @@ const fetchCache = isBrowser()
       get: (key) => sessionStorage.getItem(key),
       remove: (key) => sessionStorage.removeItem(key),
       put: (key, value) => sessionStorage.setItem(key, value),
-      clear: () => sessionStorage.clear(),
     }
   : {
       get: (key) => cache[key],
@@ -15,7 +14,6 @@ const fetchCache = isBrowser()
       put: (key, value) => {
         cache[key] = value;
       },
-      clear: () => {},
     };
 
 export const getKey = (url, opts) => url + JSON.stringify(opts);
@@ -30,9 +28,6 @@ export const writeCacheSafe = (key, value) => {
   try {
     fetchCache.put(key, value);
   } catch (e) {
-    if (e.message.includes('exceeded the quota')) {
-      fetchCache.clear();
-    }
     console.warn(`Item ${key} was not saved to cache: `, e); // eslint-disable-line no-console
   }
 };
