+++ src/utils.ts
@@ -53,12 +53,3 @@ export const getUtfStrikethrough = (text) =>
     .join('');
 
 export const join = (a, sep, b) => `${a || ''}${a && b ? sep : ''}${b || ''}`;
-
-export const publishDbgObject = (key, value) => {
-  if (typeof window !== 'undefined') {
-    // @ts-ignore
-    if (!window.dbg) window.dbg = {};
-    // @ts-ignore
-    window.dbg[key] = value;
-  }
-};
